+++ repository/book_repository.go
@@ -105,35 +105,6 @@ type CreateBookResponse struct {
 	ISBN          *string
 }
 
-// FIXME:移動
-func newInt64(ni sql.NullInt64) *int64 {
-	if !ni.Valid {
-		return nil
-	}
-	return &ni.Int64
-}
-
-// FIXME:移動
-func newString(ns sql.NullString) *string {
-	if !ns.Valid {
-		return nil
-	}
-	return &ns.String
-}
-
-// FIXME:移動
-func newTime(nt sql.NullTime) *time.Time {
-	if !nt.Valid {
-		return nil
-	}
-	return &nt.Time
-}
-
-// FIXME:移動
-func newGenres(bytes []byte) []string {
-	return strings.Split(string(bytes), ", ")
-}
-
 func newCreateResponse(b sqlc.Book) *CreateBookResponse {
 	return &CreateBookResponse{
 		ID:            b.ID,
